+++ src/Tree.Compare/Tree.Compare.psm1
@@ -14,6 +14,7 @@ Export-ModuleMember -Function @(
     'Resolve-TcEffectiveRoot'
     'New-TcFileDictionary'
     'Get-TcFileSha256'
+    'ConvertFrom-TcRobocopyLog'
     'Compare-TcTree'
     'Get-TcCompareExitCode'
     'Format-TcAgentSummary'